Add SingleBlockInventory class to get rid of boilerplate code
